+++ ludwig/combiners/combiners.py
@@ -16,21 +16,26 @@
 import logging
 from abc import ABC
 from functools import lru_cache
-from typing import Any, Dict, List, Optional, Union
+from typing import Dict
 
 import torch
-from marshmallow_dataclass import dataclass
 from torch.nn import Linear, ModuleList
 
+from ludwig.combiners.utils import sequence_encoder_registry
 from ludwig.constants import BINARY, NUMBER
-from ludwig.encoders.sequence_encoders import ParallelCNN, StackedCNN, StackedCNNRNN, StackedParallelCNN, StackedRNN
 from ludwig.features.base_feature import InputFeature
 from ludwig.modules.attention_modules import TransformerStack
 from ludwig.modules.embedding_modules import Embed
 from ludwig.modules.fully_connected_modules import FCStack
 from ludwig.modules.reduction_modules import SequenceReducer
 from ludwig.modules.tabnet_modules import TabNet
-from ludwig.schema import utils
+from ludwig.schema.combiners.comparator_combiner_config import ComparatorCombinerConfig
+from ludwig.schema.combiners.concat_combiner_config import ConcatCombinerConfig
+from ludwig.schema.combiners.sequence_combiner_config import SequenceCombinerConfig
+from ludwig.schema.combiners.sequence_concat_combiner_config import SequenceConcatCombinerConfig
+from ludwig.schema.combiners.tab_transformer_combiner_config import TabTransformerCombinerConfig
+from ludwig.schema.combiners.tabnet_combiner_config import TabNetCombinerConfig
+from ludwig.schema.combiners.transformer_combiner_config import TransformerCombinerConfig
 from ludwig.utils.misc_utils import get_from_registry
 from ludwig.utils.registry import Registry
 from ludwig.utils.torch_utils import LudwigModule, sequence_length_3D
@@ -38,15 +43,6 @@ from ludwig.utils.torch_utils import sequence_mask as torch_sequence_mask
 
 logger = logging.getLogger(__name__)
 
-sequence_encoder_registry = {
-    "stacked_cnn": StackedCNN,
-    "parallel_cnn": ParallelCNN,
-    "stacked_parallel_cnn": StackedParallelCNN,
-    "rnn": StackedRNN,
-    "cnnrnn": StackedCNNRNN,
-    # todo: add transformer
-    # 'transformer': StackedTransformer,
-}
 
 combiner_registry = Registry()
 
@@ -59,33 +55,13 @@ def register_combiner(name: str):
     return wrap
 
 
-def get_combiner_jsonschema():
-    """Returns a JSON schema structured to only require a `type` key and then conditionally apply a corresponding
-    combiner's field constraints."""
-    combiner_types = sorted(list(combiner_registry.keys()))
-    return {
-        "type": "object",
-        "properties": {
-            "type": {"type": "string", "enum": combiner_types, "default": "concat"},
-        },
-        "allOf": get_combiner_conds(),
-        "required": ["type"],
-    }
+def get_combiner_class(combiner_type: str):
+    """Returns the corresponding combiner class from `ludwig.combiners.combiners.combiner_registry`.
 
-
-def get_combiner_conds():
-    """Returns a list of if-then JSON clauses for each combiner type in `combiner_registry` and its properties'
-    constraints."""
-    combiner_types = sorted(list(combiner_registry.keys()))
-    conds = []
-    for combiner_type in combiner_types:
-        combiner_cls = combiner_registry[combiner_type]
-        schema_cls = combiner_cls.get_schema_cls()
-        combiner_schema = utils.unload_jsonschema_from_marshmallow_class(schema_cls)
-        combiner_props = combiner_schema["properties"]
-        combiner_cond = utils.create_cond({"type": combiner_type}, combiner_props)
-        conds.append(combiner_cond)
-    return conds
+    :param combiner_type: identifier that should correspond to a registered combiner
+    :type combiner_type: str
+    """
+    return get_from_registry(combiner_type, combiner_registry)
 
 
 # super class to house common properties
@@ -120,51 +96,6 @@ class Combiner(LudwigModule, ABC):
             }
         output_tensor = self.forward(pseudo_input)
         return output_tensor["combiner_output"].size()[1:]
-
-
-class BaseCombinerConfig(utils.BaseMarshmallowConfig):
-    """Base combiner config class."""
-
-    pass
-
-
-@dataclass
-class ConcatCombinerConfig(BaseCombinerConfig):
-    """Parameters for concat combiner."""
-
-    fc_layers: Optional[List[Dict[str, Any]]] = utils.DictList(description="TODO: Document parameters.")
-
-    num_fc_layers: int = utils.NonNegativeInteger(default=0, description="TODO: Document parameters.")
-
-    output_size: int = utils.PositiveInteger(default=256, description="Output size of a fully connected layer.")
-
-    use_bias: bool = utils.Boolean(default=True, description="Whether the layer uses a bias vector.")
-
-    weights_initializer: Union[str, Dict] = utils.InitializerOrDict(
-        default="xavier_uniform", description="TODO: Document parameters."
-    )
-
-    bias_initializer: Union[str, Dict] = utils.InitializerOrDict(
-        default="zeros", description="TODO: Document parameters."
-    )
-
-    norm: Optional[str] = utils.StringOptions(["batch", "layer"], description="TODO: Document parameters.")
-
-    norm_params: Optional[dict] = utils.Dict(description="TODO: Document parameters.")
-
-    activation: str = utils.ActivationOptions(default="relu", description="TODO: Document parameters.")
-
-    dropout: float = utils.FloatRange(default=0.0, min=0, max=1, description="TODO: Document parameters.")
-
-    flatten_inputs: bool = utils.Boolean(default=False, description="Whether to flatten input tensors to a vector.")
-
-    residual: bool = utils.Boolean(
-        default=False,
-        description=(
-            "Whether to add a residual connection to each fully connected layer block. All fully connected layers must"
-            " have the same size"
-        ),
-    )
 
 
 @register_combiner(name="concat")
@@ -239,15 +170,6 @@ class ConcatCombiner(Combiner):
     @staticmethod
     def get_schema_cls():
         return ConcatCombinerConfig
-
-
-@dataclass
-class SequenceConcatCombinerConfig(BaseCombinerConfig):
-    """Parameters for sequence concat combiner."""
-
-    main_sequence_feature: Optional[str] = utils.String(default=None, description="TODO: Document parameters.")
-
-    reduce_output: Optional[str] = utils.ReductionOptions(default=None, description="TODO: Document parameters.")
 
 
 @register_combiner(name="sequence_concat")
@@ -396,19 +318,6 @@ class SequenceConcatCombiner(Combiner):
         return SequenceConcatCombinerConfig
 
 
-@dataclass
-class SequenceCombinerConfig(BaseCombinerConfig):
-    """Parameters for sequence combiner."""
-
-    main_sequence_feature: Optional[str] = utils.String(default=None, description="TODO: Document parameters.")
-
-    reduce_output: Optional[str] = utils.ReductionOptions(default=None, description="TODO: Document parameters.")
-
-    encoder: Optional[str] = utils.StringOptions(
-        list(sequence_encoder_registry.keys()), default=None, description="TODO: Document parameters."
-    )
-
-
 @register_combiner(name="sequence")
 class SequenceCombiner(Combiner):
     def __init__(self, input_features: Dict[str, "InputFeature"], config: SequenceCombinerConfig = None, **kwargs):
@@ -472,68 +381,6 @@ class SequenceCombiner(Combiner):
     @staticmethod
     def get_schema_cls():
         return SequenceCombinerConfig
-
-
-@dataclass
-class TabNetCombinerConfig(BaseCombinerConfig):
-    """Parameters for tabnet combiner."""
-
-    size: int = utils.PositiveInteger(default=32, description="`N_a` in the paper.")
-
-    output_size: int = utils.PositiveInteger(
-        default=32, description="Output size of a fully connected layer. `N_d` in the paper"
-    )
-
-    num_steps: int = utils.NonNegativeInteger(
-        default=1,
-        description=(
-            "Number of steps / repetitions of the the attentive transformer and feature transformer computations. "
-            "`N_steps` in the paper"
-        ),
-    )
-
-    num_total_blocks: int = utils.NonNegativeInteger(
-        default=4, description="Total number of feature transformer block at each step"
-    )
-
-    num_shared_blocks: int = utils.NonNegativeInteger(
-        default=2, description="Number of shared feature transformer blocks across the steps"
-    )
-
-    relaxation_factor: float = utils.FloatRange(
-        default=1.5,
-        description=(
-            "Factor that influences how many times a feature should be used across the steps of computation. a value of"
-            " 1 implies it each feature should be use once, a higher value allows for multiple usages. `gamma` in the "
-            "paper"
-        ),
-    )
-
-    bn_epsilon: float = utils.FloatRange(default=1e-3, description="Epsilon to be added to the batch norm denominator.")
-
-    bn_momentum: float = utils.FloatRange(default=0.7, description="Momentum of the batch norm. `m_B` in the paper.")
-
-    bn_virtual_bs: Optional[int] = utils.PositiveInteger(
-        default=None,
-        description=(
-            "Size of the virtual batch size used by ghost batch norm. If null, regular batch norm is used instead. "
-            "`B_v` from the paper"
-        ),
-    )
-
-    sparsity: float = utils.FloatRange(
-        default=1e-5, description="Multiplier of the sparsity inducing loss. `lambda_sparse` in the paper"
-    )
-
-    entmax_mode: str = utils.StringOptions(
-        ["entmax15", "sparsemax", "constant", "adaptive"], default="sparsemax", description="TODO: Document parameters."
-    )
-
-    entmax_alpha: float = utils.FloatRange(
-        default=1.5, min=1, max=2, description="TODO: Document parameters."
-    )  # 1 corresponds to softmax, 2 is sparsemax.
-
-    dropout: float = utils.FloatRange(default=0.0, min=0, max=1, description="Dropout rate for the transformer block.")
 
 
 @register_combiner(name="tabnet")
@@ -614,73 +461,6 @@ class TabNetCombiner(Combiner):
     @property
     def output_shape(self) -> torch.Size:
         return self.tabnet.output_shape
-
-
-@dataclass
-class CommonTransformerConfig:
-    """Common transformer parameter values."""
-
-    num_layers: int = utils.PositiveInteger(default=1, description="TODO: Document parameters.")
-
-    hidden_size: int = utils.NonNegativeInteger(
-        default=256,
-        description=(
-            "The number of hidden units of the TransformerStack as well as the dimension that each incoming input "
-            "feature is projected to before feeding to the TransformerStack"
-        ),
-    )
-
-    num_heads: int = utils.NonNegativeInteger(
-        default=8, description="Number of heads of the self attention in the transformer block."
-    )
-
-    transformer_output_size: int = utils.NonNegativeInteger(
-        default=256,
-        description=(
-            "Size of the fully connected layer after self attention in the transformer block. This is usually the same "
-            "as `hidden_size` and `embedding_size`."
-        ),
-    )
-
-    dropout: float = utils.FloatRange(default=0.1, min=0, max=1, description="Dropout rate for the transformer block.")
-
-    fc_layers: Optional[List[Dict[str, Any]]] = utils.DictList(description="TODO: Document parameters.")
-
-    # TODO(#1673): Add conditional logic for fields like this one:
-    num_fc_layers: int = utils.NonNegativeInteger(
-        default=0,
-        description="The number of stacked fully connected layers (only applies if `reduce_output` is not null).",
-    )
-
-    output_size: int = utils.PositiveInteger(default=256, description="Output size of a fully connected layer.")
-
-    use_bias: bool = utils.Boolean(default=True, description="Whether the layer uses a bias vector.")
-
-    weights_initializer: Union[str, Dict] = utils.InitializerOrDict(
-        default="xavier_uniform", description="TODO: Document parameters."
-    )
-
-    bias_initializer: Union[str, Dict] = utils.InitializerOrDict(
-        default="zeros", description="TODO: Document parameters."
-    )
-
-    norm: Optional[str] = utils.StringOptions(["batch", "layer"], description="TODO: Document parameters.")
-
-    norm_params: Optional[dict] = utils.Dict(description="TODO: Document parameters.")
-
-    fc_activation: str = utils.ActivationOptions(default="relu", description="TODO: Document parameters.")
-
-    fc_dropout: float = utils.FloatRange(default=0.0, min=0, max=1, description="TODO: Document parameters.")
-
-    fc_residual: bool = utils.Boolean(default=False, description="TODO: Document parameters.")
-
-
-# TODO(ksbrar): Refactor transformers into using base class for common attrs?
-@dataclass
-class TransformerCombinerConfig(BaseCombinerConfig, CommonTransformerConfig):
-    """Parameters for transformer combiner."""
-
-    reduce_output: Optional[str] = utils.ReductionOptions(default="mean", description="TODO: Document parameters.")
 
 
 @register_combiner(name="transformer")
@@ -781,15 +561,6 @@ class TransformerCombiner(Combiner):
     @staticmethod
     def get_schema_cls():
         return TransformerCombinerConfig
-
-
-@dataclass
-class TabTransformerCombinerConfig(BaseCombinerConfig, CommonTransformerConfig):
-    """Parameters for tab transformer combiner."""
-
-    embed_input_feature_name: Optional[Union[str, int]] = utils.Embed()
-
-    reduce_output: str = utils.ReductionOptions(default="concat", description="TODO: Document parameters.")
 
 
 @register_combiner(name="tabtransformer")
@@ -996,41 +767,6 @@ class TabTransformerCombiner(Combiner):
         return TabTransformerCombinerConfig
 
 
-@dataclass
-class ComparatorCombinerConfig(BaseCombinerConfig):
-    """Parameters for comparator combiner."""
-
-    entity_1: List[str]
-    """TODO: Document parameters."""
-
-    entity_2: List[str]
-    """TODO: Document parameters."""
-
-    fc_layers: Optional[List[Dict[str, Any]]] = utils.DictList(description="TODO: Document parameters.")
-
-    num_fc_layers: int = utils.NonNegativeInteger(default=1, description="TODO: Document parameters.")
-
-    output_size: int = utils.PositiveInteger(default=256, description="Output size of a fully connected layer")
-
-    use_bias: bool = utils.Boolean(default=True, description="Whether the layer uses a bias vector.")
-
-    weights_initializer: Union[str, Dict] = utils.InitializerOrDict(
-        default="xavier_uniform", description="TODO: Document parameters."
-    )
-
-    bias_initializer: Union[str, Dict] = utils.InitializerOrDict(
-        default="zeros", description="TODO: Document parameters."
-    )
-
-    norm: Optional[str] = utils.StringOptions(["batch", "layer"], description="TODO: Document parameters.")
-
-    norm_params: Optional[dict] = utils.Dict(description="TODO: Document parameters.")
-
-    activation: str = utils.ActivationOptions(default="relu", description="TODO: Document parameters.")
-
-    dropout: float = utils.FloatRange(default=0.0, min=0, max=1, description="Dropout rate for the transformer block.")
-
-
 @register_combiner(name="comparator")
 class ComparatorCombiner(Combiner):
     def __init__(
@@ -1176,12 +912,3 @@ class ComparatorCombiner(Combiner):
     @staticmethod
     def get_schema_cls():
         return ComparatorCombinerConfig
-
-
-def get_combiner_class(combiner_type: str):
-    """Returns the corresponding combiner class from `ludwig.combiners.combiners.combiner_registry`.
-
-    :param combiner_type: identifier that should correspond to a registered combiner
-    :type combiner_type: str
-    """
-    return get_from_registry(combiner_type, combiner_registry)
