+++ ludwig/schema/combiners/base_combiner_config.py
@@ -1,0 +1,7 @@
+from ludwig.schema import utils
+
+
+class BaseCombinerConfig(utils.BaseMarshmallowConfig):
+    """Base combiner config class."""
+
+    pass
