+++ ludwig/schema/combiners/utils.py
@@ -1,0 +1,31 @@
+from ludwig.combiners.combiners import combiner_registry
+from ludwig.schema import utils
+
+
+def get_combiner_jsonschema():
+    """Returns a JSON schema structured to only require a `type` key and then conditionally apply a corresponding
+    combiner's field constraints."""
+    combiner_types = sorted(list(combiner_registry.keys()))
+    return {
+        "type": "object",
+        "properties": {
+            "type": {"type": "string", "enum": combiner_types, "default": "concat"},
+        },
+        "allOf": get_combiner_conds(),
+        "required": ["type"],
+    }
+
+
+def get_combiner_conds():
+    """Returns a list of if-then JSON clauses for each combiner type in `combiner_registry` and its properties'
+    constraints."""
+    combiner_types = sorted(list(combiner_registry.keys()))
+    conds = []
+    for combiner_type in combiner_types:
+        combiner_cls = combiner_registry[combiner_type]
+        schema_cls = combiner_cls.get_schema_cls()
+        combiner_schema = utils.unload_jsonschema_from_marshmallow_class(schema_cls)
+        combiner_props = combiner_schema["properties"]
+        combiner_cond = utils.create_cond({"type": combiner_type}, combiner_props)
+        conds.append(combiner_cond)
+    return conds
